requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

fastapi==0.99.1
loguru==0.7.2
python-dotenv==1.0.0
requests==2.28.2
uvicorn==0.23.2
